% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.element_attr_df}
\alias{.element_attr_df}
\title{Get XML attributes}
\usage{
.element_attr_df(c1, c2, dat)
}
\arguments{
\item{c1}{numeric: xml2 child element}

\item{c2}{numeric: xml2 child element}

\item{dat}{xml document}
}
\value{
data frame of xml attributes
}
\description{
This helper function extracts xml elements specific to CXL docs. It takes two
child element arguments, as is needed for CXL document traversing
}
